Removed `send_ref_offset_value' which hasen't been used for awhile. Fixed various C compiler warnings raised by clang. Now the runtime compiles without warnings with -Wall and -Wextra, but there are still quite a bit of warnings if using -Weverything: * Change many types from numeral int/long to size_t which in many cases enable us to get rid of casts (that are always hiding some issues). * Change some numeral to `uint32' to match what is expected (but maybe size_t would be better, but this is a more complicated change). * Added new header `ewb_init.h' to provide a prototype to public routines of `ewb_init.c'. * Fixed missing declarations, unused macros, ... * Changed declaration of EIF_process_once_count to be size_t and avoid many warnings.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@95924 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/ipc/app/app_transfer.c
@@ -71,7 +71,7 @@ rt_public void app_tpipe(STREAM *stream)
 #endif
 }
 
-rt_public char *app_tread(int *size)
+rt_public char *app_tread(size_t *size)
           		/* Filled in with size of read string */
 {
 	/* Read bytes from the "pipe" and put them into a new allocated buffer.
